// employee-management-system/src/app/models/project.ts
export interface Project {
    id: string;
    date: string;
    tower: string;
    techPod: string;
    solution: string;
    status: string;
    domainArchitecture: string;
    productOwner: string;
    scrumMaster: string;
    uxDesigner: string;
    technicalLead: string;
    developer1: string;
    developer2: string;
    developer3: string;
    developer4: string;
    developer5: string;
    developer6: string;
    developer7: string;
}